requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Scikit-Learn==1.5.0
Pandas==2.2.0
Numpy==1.26.0
matplotlib==3.9.0
Seaborn==0.13.0
PyMongo==4.6.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.1.2
Flask-SQLAlchemy==3.1.1